// src/app/models/User.ts
export class User {
    uid: number | undefined;
    firstName: String | undefined;
    surname: String | undefined;
    identify: String | undefined;
    password: String | undefined;
    definition: String | undefined;
    phone: String | undefined;
    token: string | undefined;
    ip: String | undefined;


    constructor(){};
}